chore: refactor BulkCommitBatch back in to extend WriteBatch
