+++ dev/test/write-batch.ts
@@ -435,93 +435,3 @@ describe('batch support', () => {
     });
   });
 });
-
-describe('bulkCommit support', () => {
-  const doc1 = `projects/${PROJECT_ID}/databases/(default)/documents/col/doc1`;
-  const doc2 = `projects/${PROJECT_ID}/databases/(default)/documents/col/doc2`;
-
-  let firestore: Firestore;
-  let writeBatch: WriteBatch;
-
-  beforeEach(() => {
-    const overrides: ApiOverride = {
-      batchWrite: request => {
-        expect(request).to.deep.eq({
-          database: `projects/${PROJECT_ID}/databases/(default)`,
-          writes: [
-            {
-              update: {
-                fields: {},
-                name: doc1,
-              },
-              updateTransforms: [
-                {
-                  fieldPath: 'foo',
-                  setToServerValue: REQUEST_TIME,
-                },
-              ],
-            },
-            {
-              currentDocument: {
-                exists: true,
-              },
-              update: {
-                fields: {
-                  foo: {
-                    stringValue: 'bar',
-                  },
-                },
-                name: doc2,
-              },
-              updateMask: {
-                fieldPaths: ['foo'],
-              },
-            },
-          ],
-        });
-        return response({
-          writeResults: [
-            {
-              updateTime: {
-                nanos: 0,
-                seconds: 0,
-              },
-            },
-            {
-              updateTime: null,
-            },
-          ],
-          status: [{code: 0}, {code: 4}],
-        });
-      },
-    };
-    return createInstance(overrides).then(firestoreClient => {
-      firestore = firestoreClient;
-      writeBatch = firestore.batch();
-    });
-  });
-
-  afterEach(() => verifyInstance(firestore));
-
-  it('bulkCommit', () => {
-    const doc1 = firestore.doc('col/doc1');
-    const doc2 = firestore.doc('col/doc2');
-    const completed: string[] = [];
-
-    writeBatch.set(doc1, {foo: FieldValue.serverTimestamp()});
-    writeBatch.update(doc2, {foo: 'bar'});
-    writeBatch._processLastOperation(doc1).then(res => {
-      expect(res.writeTime!.isEqual(new Timestamp(0, 0))).to.be.true;
-      completed.push('doc1');
-    });
-    writeBatch._processLastOperation(doc2).catch(err => {
-      expect(err.code).to.equal(Status.DEADLINE_EXCEEDED);
-      completed.push('doc2');
-    });
-    writeBatch._markReadyToSend();
-    return writeBatch.bulkCommit().then(() => {
-      expect(writeBatch._pendingOps.length).to.equal(2);
-      expect(completed).to.deep.equal(['doc1', 'doc2']);
-    });
-  });
-});
