+++ package.json
@@ -54,7 +54,8 @@
   "dependencies": {
     "fast-deep-equal": "^3.1.1",
     "functional-red-black-tree": "^1.0.1",
-    "google-gax": "^2.2.0"
+    "google-gax": "^2.2.0",
+    "madge": "^3.12.0"
   },
   "devDependencies": {
     "@types/assert": "^1.4.0",
